+++ pg_lakehouse/src/datafusion/mod.rs
@@ -1,4 +1,5 @@
 pub mod context;
+pub mod format;
 pub mod plan;
 pub mod provider;
 pub mod session;
