+++ pg_lakehouse/src/fdw/mod.rs
@@ -1,3 +1,2 @@
-pub mod format;
 pub mod handler;
 pub mod options;
